move AnyToNbtSerializer to common
