+++ serialization/src/test/kotlin/io/github/leaflowmc/leaflow/serialization/test/AnyToNbtSerializerTest.kt
@@ -3,7 +3,7 @@ package io.github.leaflowmc.leaflow.serialization.test
 import io.github.leaflowmc.leaflow.commonTest.utils.byteBufBytes
 import io.github.leaflowmc.leaflow.serialization.minecraft_format.decode
 import io.github.leaflowmc.leaflow.serialization.minecraft_format.encode
-import io.github.leaflowmc.leaflow.serialization.nbt.AnyToNbtSerializer
+import io.github.leaflowmc.leaflow.common.serializer.AnyToNbtSerializer
 import io.netty.buffer.Unpooled
 import kotlinx.serialization.Serializable
 import net.kyori.adventure.nbt.BinaryTagIO
